moving stuff around a bit, trying to get rid of types
